move math plugin to com.fossgalaxy namespace
