+++ src/main/java/com/fossgalaxy/pircbotx/commands/math/MathFunction.java
@@ -1,4 +1,4 @@
-package uk.co.unitycoders.pircbotx.commands.math;
+package com.fossgalaxy.pircbotx.commands.math;
 
 public class MathFunction implements Function {
 	private String name;
