+++ src/main/java/com/fossgalaxy/pircbotx/commands/math/OperatorFunction.java
@@ -1,4 +1,4 @@
-package uk.co.unitycoders.pircbotx.commands.math;
+package com.fossgalaxy.pircbotx.commands.math;
 
 public class OperatorFunction implements Function {
 	private String name;
