+++ src/main/java/com/fossgalaxy/pircbotx/commands/math/PostfixEvaulator.java
@@ -1,4 +1,4 @@
-package uk.co.unitycoders.pircbotx.commands.math;
+package com.fossgalaxy.pircbotx.commands.math;
 
 import java.util.Map;
 import java.util.Queue;
