+++ src/main/java/com/fossgalaxy/pircbotx/commands/math/Token.java
@@ -1,4 +1,4 @@
-package uk.co.unitycoders.pircbotx.commands.math;
+package com.fossgalaxy.pircbotx.commands.math;
 
 public class Token {
 	private TokenType type;
