+++ src/main/java/com/fossgalaxy/pircbotx/commands/math/TokenType.java
@@ -1,4 +1,4 @@
-package uk.co.unitycoders.pircbotx.commands.math;
+package com.fossgalaxy.pircbotx.commands.math;
 
 public enum TokenType {
 	NUMBER,
