+++ src/main/java/com/fossgalaxy/pircbotx/commands/math/Tokeniser.java
@@ -1,4 +1,4 @@
-package uk.co.unitycoders.pircbotx.commands.math;
+package com.fossgalaxy.pircbotx.commands.math;
 
 import java.util.LinkedList;
 import java.util.Queue;
