+++ src/main/java/com/fossgalaxy/pircbotx/commands/math/TreeNode.java
@@ -1,4 +1,4 @@
-package uk.co.unitycoders.pircbotx.commands.math;
+package com.fossgalaxy.pircbotx.commands.math;
 
 import java.util.Arrays;
 import java.util.Collections;
