+++ src/main/resources/META-INF/services/com.fossgalaxy.pircbotx.modules.Module
@@ -10,4 +10,4 @@ com.fossgalaxy.pircbotx.commands.KarmaCommand
 com.fossgalaxy.pircbotx.commands.GamesCommand
 com.fossgalaxy.pircbotx.commands.CompSciCommand
 com.fossgalaxy.pircbotx.commands.IRCCommands
-uk.co.unitycoders.pircbotx.commands.math.MathCommand
+com.fossgalaxy.pircbotx.commands.math.MathCommand
