+++ src/test/java/com/fossgalaxy/pircbotx/commands/math/BasicMathTests.java
@@ -1,4 +1,4 @@
-package uk.co.unitycoders.pircbotx.commands.math;
+package com.fossgalaxy.pircbotx.commands.math;
 
 import org.junit.Assert;
 import org.junit.Before;
